Deprecate pretty printer, update to use mbed-trace
This commit deprecates the functions in pretty printer and replaces them with new versions in ble_logging that use the standard mbed-trace library for BLE utility logging.

A library should not use printf when a trace library is available that allows you to configure/enable/disable output. This prevents the debug output from being cluttered with unformatted debug messages from the BLE utility library.